requirements.txt
streamlit==1.33.0
pandas==2.2.2
langgraph==0.0.35
langchain==0.1.14
langchain-community==0.0.27
openai==1.16.1
python-dotenv==1.0.1
pydantic<2.6
